GetLogProdScales endColumn should be unsigned
The only time ScaledMatrix::GetLogProdScales gets called with an int as the second argument is in EvaulatorImpl.cpp::128; since mut->Start() is >= 3, the -1 will never cause extendStartCol to be negative.

+++ src/matrix/ScaledMatrix.h
@@ -73,7 +73,7 @@ public:  // information about entries filled by column
 
 public:  // Scaling and normalization
     double GetLogScale(size_t j) const;
-    double GetLogProdScales(size_t s, int e) const;
+    double GetLogProdScales(size_t s, size_t e) const;
     double GetLogProdScales() const;
 
 public:  // Convenient matrix access for SWIG
@@ -125,7 +125,7 @@ inline void ScaledMatrix::FinishEditingColumn(const size_t j, const size_t usedB
 }
 
 inline double ScaledMatrix::GetLogScale(size_t j) const { return logScalars_[j]; }
-inline double ScaledMatrix::GetLogProdScales(size_t beginColumn, int endColumn) const
+inline double ScaledMatrix::GetLogProdScales(size_t beginColumn, size_t endColumn) const
 {
     double f, l;
     if (dir_ == FORWARD) {
